% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comarm.R
\name{comarm}
\alias{comarm}
\title{Fit composed model (CoMARM) with sparsity assumption and fixed ranks.}
\usage{
comarm(
  Y,
  X1,
  X2,
  G1 = NULL,
  group = NULL,
  is.fabs = 1,
  K = 6,
  r31 = NULL,
  r32 = NULL,
  r33 = NULL,
  r41 = NULL,
  r42 = NULL,
  r43 = NULL,
  r44 = NULL,
  method = "BIC",
  ncv = 10,
  penalty = "LASSO",
  lambda = NULL,
  D0_t3 = NULL,
  D0_t4 = NULL,
  intercept = TRUE,
  nlam = 50,
  degr = 3,
  lam_min = 0.01,
  eps = 1e-04,
  max_step = 20,
  eps1 = 1e-04,
  max_step1_t3 = 20,
  max_step1_t4 = 20,
  gamma = 2,
  dfmax1 = NULL,
  dfmax2 = NULL,
  alpha = 1,
  vnorm_ratio = 1
)
}
\arguments{
\item{Y}{A \eqn{n \times q} numeric matrix of responses.}

\item{X1}{A \eqn{n \times p1} numeric design matrix for the first part of the model.}

\item{X2}{A \eqn{n \times p2} numeric design matrix for the second part of the model.}

\item{G1}{Number of views without intergroup correlation.}

\item{group}{Grouping index for predictors.}

\item{is.fabs}{Logical indicating if data comes from scenario I.}

\item{K}{Number of B-spline basis functions.}

\item{r31}{First dimension rank of third-order tensor.}

\item{r32}{Second dimension rank of third-order tensor.}

\item{r33}{Third dimension rank of third-order tensor.}

\item{r41}{First dimension rank of fourth-order tensor.}

\item{r42}{Second dimension rank of fourth-order tensor.}

\item{r43}{Third dimension rank of fourth-order tensor.}

\item{r44}{Fourth dimension rank of fourth-order tensor.}

\item{method}{Selection method for model parameters (BIC or CV).}

\item{ncv}{Number of cross-validation folds.}

\item{penalty}{Type of penalty (LASSO, MCP, SCAD).}

\item{lambda}{Sequence of lambda values for regularization.}

\item{D0_t3}{Initial values for third-order tensor decomposition.}

\item{D0_t4}{Initial values for fourth-order tensor decomposition.}

\item{intercept}{Whether to include an intercept.}

\item{nlam}{Number of lambda values.}

\item{degr}{Number of knots in the B-spline.}

\item{lam_min}{Minimum lambda as a fraction of the maximum.}

\item{eps}{Convergence threshold.}

\item{max_step}{Maximum number of iterations allowed.}

\item{eps1}{Convergence threshold for coordinate descent.}

\item{max_step1_t3}{Maximum iterations for coordinate descent on third-order tensor.}

\item{max_step1_t4}{Maximum iterations for coordinate descent on fourth-order tensor.}

\item{gamma}{Tuning parameter for MCP or SCAD.}

\item{dfmax1}{Maximum number of non-zero coefficients for third-order tensor.}

\item{dfmax2}{Maximum number of non-zero coefficients for fourth-order tensor.}

\item{alpha}{Balance parameter between LASSO, MCP/SCAD, and ridge.}

\item{vnorm_ratio}{Ratio between lambda values for different tensor types.}
}
\value{
A list containing model estimates, diagnostic measures, and tensor components.
}
\description{
Fits a composed model for multi-view data using B-splines with specified ranks 
(\eqn{r_{1g}, r_{2g}, r_{3g}} and \eqn{r_{1}, r_{2}, r_{3}, r_{4}}). This function estimates
multiple third-order and a single fourth-order coefficient tensors, applying group sparse
penalties such as LASSO, MCP, or SCAD along with a coordinate descent algorithm to optimize
the sparsity estimator. The BIC or cross-validation methods are used to select the optimal
regularization parameters.
}
\examples{
library(comarm)
n <- 200; q <- 5; p <- 100; s1 <- 5; s2 <- 3; G1 <- 1; ng <- 4
group <- rep(1:ng, each = p/ng)
mydata <- comarm.sim.fbs(n, q, p, s1, s2, G1, group, isfixedR = 1)
fit <- with(mydata, comarm(Y, X1, X2, G1, group, is.fabs = 1, K = 6, 
              r31 = 2, r32 = 2, r33 = 2, r41 = 2, r42 = 2, r43 = 2, r44 = 2, 
              method = "BIC", ncv = 10, penalty = "LASSO", lambda = NULL, 
              D0_t3 = NULL, D0_t4 = NULL, intercept = TRUE, nlam = 50, degr = 3, 
              lam_min = 0.01, eps = 1e-4, max_step = 20, eps1 = 1e-4, 
              max_step1_t3 = 20, max_step1_t4 = 20, gamma = 2, dfmax1 = NULL, 
              dfmax2 = NULL, alpha = 1, vnorm_ratio = 1))

}
\seealso{
\code{\link{comarm.dr}}
}
