% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comarm.sim.fbs.R
\name{comarm.sim.fbs}
\alias{comarm.sim.fbs}
\title{Generate scenario I data from composed model (CoMARM).}
\usage{
comarm.sim.fbs(
  n,
  q,
  p,
  s1,
  s2,
  G1 = NULL,
  group = NULL,
  r310 = 2,
  r320 = 2,
  r330 = 2,
  r410 = 2,
  r420 = 2,
  r430 = 2,
  r440 = 2,
  isfixedR = 0,
  D3 = NULL,
  D44 = NULL,
  K = 6,
  degr = 3,
  sigma2 = NULL,
  seed_id = NULL,
  r1_t3_index = NULL,
  r2_t3_index = NULL,
  r3_t3_index = NULL,
  r1_t4_index = NULL,
  r2_t4_index = NULL,
  r3_t4_index = NULL,
  r4_t4_index = NULL,
  D0_t3 = NULL,
  D0_t4 = NULL
)
}
\arguments{
\item{n}{Sample size.}

\item{q}{Number of responses.}

\item{p}{Total number of covariates.}

\item{s1}{Number of true covariates associated with the first set of views.}

\item{s2}{Number of true covariates associated with the second set of views.}

\item{G1}{Number of views without intergroup correlations.}

\item{group}{Grouping index for predictors.}

\item{r310}{First rank dimension for third-order tensors.}

\item{r320}{Second rank dimension for third-order tensors.}

\item{r330}{Third rank dimension for third-order tensors.}

\item{r410}{First rank dimension for fourth-order tensors.}

\item{r420}{Second rank dimension for fourth-order tensors.}

\item{r430}{Third rank dimension for fourth-order tensors.}

\item{r440}{Fourth rank dimension for fourth-order tensors.}

\item{isfixedR}{Logical indicating if ranks are fixed.}

\item{D3}{Initial data for third-order tensor simulation.}

\item{D44}{Initial data for fourth-order tensor simulation.}

\item{K}{Number of B-spline basis functions.}

\item{degr}{Number of knots in B-spline.}

\item{sigma2}{Error variance.}

\item{seed_id}{Seed for random number generation.}

\item{r1_t3_index}{Rank indices for the first dimension of third-order tensors.}

\item{r2_t3_index}{Rank indices for the second dimension of third-order tensors.}

\item{r3_t3_index}{Rank indices for the third dimension of third-order tensors.}

\item{r1_t4_index}{Rank indices for the first dimension of fourth-order tensors.}

\item{r2_t4_index}{Rank indices for the second dimension of fourth-order tensors.}

\item{r3_t4_index}{Rank indices for the third dimension of fourth-order tensors.}

\item{r4_t4_index}{Rank indices for the fourth dimension of fourth-order tensors.}

\item{D0_t3}{Initial third-order tensor decomposition.}

\item{D0_t4}{Initial fourth-order tensor decomposition.}
}
\value{
A list with the following components:
  \item{Y}{Response matrix, dimensions \eqn{n \times q}.}
  \item{X1}{First set of design matrices, dimensions \eqn{n \times p1}.}
  \item{X2}{Second set of design matrices, dimensions \eqn{n \times p2}.}
  \item{f01}{True function values for the first set of views.}
  \item{f02}{True function values for the second set of views.}
  \item{group}{Grouping index of predictors.}
  \item{D0_t3}{Initialized values for the third-order tensor.}
  \item{D0_t4}{Initialized values for the fourth-order tensor.}
  \item{...}{Additional algorithm options.}
}
\description{
This function generates data for scenario I of a composed model (CoMARM), which uses B-splines 
to model data with known tensor ranks for multiple third-order and fourth-order coefficient tensors. 
It's suitable for simulations where ranks are fixed and known a priori, allowing detailed study 
of model behavior under controlled conditions.
}
\examples{
library(comarm)
n <- 200; q <- 5; p <- 100; s1 <- 5; s2 <- 3; G1 <- 1; ng = 4
group <- rep(1:ng, each = p/ng)
mydata <- comarm.sim.fbs(n, q, p, s1, s2, G1, group)

}
\seealso{
\code{\link{comarm.sim.fsin}}
}
