% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comarm.sim.fsin.R
\name{comarm.sim.fsin}
\alias{comarm.sim.fsin}
\title{Generate scenario II data from composed model (CoMARM).}
\usage{
comarm.sim.fsin(
  n,
  q,
  p,
  s1,
  s2,
  G1 = NULL,
  group = NULL,
  r310 = 2,
  r320 = 2,
  r330 = 2,
  r410 = 2,
  r420 = 2,
  r430 = 2,
  r440 = 2,
  isfixedR = 0,
  D2 = NULL,
  D42 = NULL,
  K = 6,
  degr = 3,
  sigma2 = NULL,
  seed_id = NULL,
  r1_t3_index = NULL,
  r2_t3_index = NULL,
  r3_t3_index = NULL,
  r1_t4_index = NULL,
  r2_t4_index = NULL,
  r3_t4_index = NULL,
  r4_t4_index = NULL,
  D0_t3 = NULL,
  D0_t4 = NULL
)
}
\arguments{
\item{n}{Sample size.}

\item{q}{Number of responses.}

\item{p}{Total number of covariates.}

\item{s1}{True covariates for the first set of views.}

\item{s2}{True covariates for the second set of views.}

\item{G1}{Number of views without intergroup correlations.}

\item{group}{Grouping index for predictors.}

\item{r310}{First dimension rank for third-order tensors.}

\item{r320}{Second dimension rank for third-order tensors.}

\item{r330}{Third dimension rank for third-order tensors.}

\item{r410}{First dimension rank for fourth-order tensors.}

\item{r420}{Second dimension rank for fourth-order tensors.}

\item{r430}{Third dimension rank for fourth-order tensors.}

\item{r440}{Fourth dimension rank for fourth-order tensors.}

\item{isfixedR}{Logical indicating if ranks are fixed.}

\item{D2}{Mode of unfolding for the D2 matrix, typically generated randomly.}

\item{D42}{Mode of unfolding for the D42 matrix, typically generated randomly.}

\item{K}{Number of B-spline basis functions, typically 6 for cubic splines.}

\item{degr}{Number of knots in B-spline base functions.}

\item{sigma2}{Error variance.}

\item{seed_id}{Seed for random number generation.}

\item{r1_t3_index}{Rank indices for the first dimension of third-order tensors.}

\item{r2_t3_index}{Rank indices for the second dimension of third-order tensors.}

\item{r3_t3_index}{Rank indices for the third dimension of third-order tensors.}

\item{r1_t4_index}{Rank indices for the first dimension of fourth-order tensors.}

\item{r2_t4_index}{Rank indices for the second dimension of fourth-order tensors.}

\item{r3_t4_index}{Rank indices for the third dimension of fourth-order tensors.}

\item{r4_t4_index}{Rank indices for the fourth dimension of fourth-order tensors.}

\item{D0_t3}{Initial values for the decomposition of third-order tensors.}

\item{D0_t4}{Initial values for the decomposition of fourth-order tensors.}
}
\value{
A list with components including the response matrix, design matrices for both sets of views,
true function matrices, grouping indices, and initial tensor decompositions.
}
\description{
This function generates data for scenario II of a composed model (CoMARM), which is designed to simulate
conditions where the true functions are linear combinations of sine and cosine functions. This setup is
particularly useful for testing the model's ability to capture complex, periodic relationships within the data.
}
\examples{
library(comarm)
n <- 200; q <- 5; p <- 100; s1 <- 5; s2 <- 3; G1 <- 1; ng = 4
group <- rep(1:ng, each = p/ng)
mydata <- comarm.sim.fsin(n, q, p, s1, s2, G1, group)

}
\seealso{
\code{\link{comarm.sim.fbs}}
}
