% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marm3.sim.fbs.R
\name{marm3.sim.fbs}
\alias{marm3.sim.fbs}
\title{Generate scenario I data from MARM model.}
\usage{
marm3.sim.fbs(
  n,
  q,
  p,
  s,
  group = NULL,
  r10 = 2,
  r20 = 2,
  r30 = 2,
  isfixedR = 0,
  D3 = NULL,
  K = 6,
  degr = 3,
  sigma2 = NULL,
  seed_id = NULL,
  r1_index = NULL,
  r2_index = NULL,
  r3_index = NULL,
  D0 = NULL
)
}
\arguments{
\item{n}{Sample size.}

\item{q}{The number of responses, \eqn{q \geq 1}.}

\item{p}{The number of covariates, \eqn{p \geq 1}.}

\item{s}{The true covariates of each view associated with responses, \eqn{s \geq 1}.}

\item{group}{A length-\eqn{p} vector of the grouping index of predictors, e.g., 
\eqn{group = c(1, 1, 1, 2, 2, 2)} means there are \eqn{6} predictors in the model, 
and the first three predictors are in the same group and the last three predictors 
are in another one. By default, set as \code{group = rep(1, p)}.}

\item{r10}{The first dimension of the tensor. Default is \code{2}.}

\item{r20}{The second dimension of the tensor. Default is \code{2}.}

\item{r30}{The third dimension of the tensor. Default is \code{2}.}

\item{isfixedR}{A logical value indicating whether ranks are fixed.}

\item{D3}{The mode of unfolding \eqn{D_{(3)}}. By default, generated randomly.}

\item{K}{The number of B-spline basis functions, which is the sum of both the degrees of basis 
functions and the number of knots. Default is \code{6}, which indicates cubic splines.}

\item{degr}{The number of knots of B-spline base function. Default is \code{3}.}

\item{sigma2}{Error variance. Default is \code{0.1}.}

\item{seed_id}{A positive integer, the seed for generating the random numbers. Default is \code{1000}.}

\item{r1_index}{A user-specified sequence of \eqn{r_1} values.
Default is \code{r1_index}\eqn{=1, \ldots, \min(\lceil\log(n)\rceil, p)}.
Ignored if \code{isfixedR = 1}.}

\item{r2_index}{A user-specified sequence of \eqn{r_2} values. Default is \code{r2_index = 1, \dots, 
max(K)}. Ignored if \code{isfixedR = 1}.}

\item{r3_index}{A user-specified sequence of \eqn{r_3} values. Default is \code{r3_index}\eqn{ = 1, \dots, 
\min(\lceil\log(n)\rceil, q)}. Ignored if \code{isfixedR = 1}.}

\item{D0}{A user-specified list of initialized values, including \code{ng} sub-lists where ng is the 
number of groups. Each sub-list has four initialized matrices \eqn{S_{(3)}} (called \code{S}), 
\code{A}, \code{B} and \code{C}. By default, a list of initialization satisfying fixed ranks is 
computed randomly.}
}
\value{
A list containing:
  \itemize{
    \item{Y}{Response, a \eqn{n \times q}-matrix.}
    \item{X}{Design matrix, a \eqn{n \times p}-matrix.}
    \item{f0}{True functions, a \eqn{n \times p}-matrix.}
    \item{group}{The grouping index of predictors, a length-\eqn{p} vector.}
    \item{D0}{The initialized values.}
    \item{...}{Other options for the algorithm.}
  }
}
\description{
Generate scenario I data for MARM model using B-splines. The function is designed
to simulate multi-view data with specified tensor ranks and a group sparsity structure.
The simulation settings allow control over the spline basis functions and the initialization
of tensor decompositions.
}
\examples{
library(comarm)
n <- 200; q <- 5; p <- 100; s <- 3; ng <- 4
group <- rep(1:ng, each = p/ng)
mydata <- marm3.sim.fbs(n, q, p, s, group)

}
\seealso{
\code{\link{marm3.sim.fsin}}
}
