% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marm3.sim.fsin.R
\name{marm3.sim.fsin}
\alias{marm3.sim.fsin}
\title{Generate scenario II data from MARM model.}
\usage{
marm3.sim.fsin(
  n,
  q,
  p,
  s,
  group = NULL,
  r10 = 2,
  r20 = 2,
  r30 = 2,
  isfixedR = 0,
  D2 = NULL,
  K = 6,
  degr = 3,
  sigma2 = NULL,
  seed_id = NULL,
  r1_index = NULL,
  r2_index = NULL,
  r3_index = NULL,
  D0 = NULL
)
}
\arguments{
\item{n}{Sample size.}

\item{q}{The number of responses, \eqn{q \geq 1}.}

\item{p}{The number of covariates, \eqn{p \geq 1}.}

\item{s}{The true covariates of each view associated with responses, \eqn{s \geq 1}.}

\item{group}{A length-\eqn{p} vector of the grouping index of predictors, typically set as \code{group = rep(1, p)}.
Default groups \eqn{group = c(1, 1, 1, 2, 2, 2)}, which partitions the predictors into two groups.}

\item{r10}{The first dimension of the tensor. Default is \code{2}.}

\item{r20}{The second dimension of the tensor. Default is \code{2}.}

\item{r30}{The third dimension of the tensor. Default is \code{2}.}

\item{isfixedR}{Logical value indicating if ranks are fixed.}

\item{D2}{Mode of unfolding \eqn{D_{(2)}}, generated randomly by default.}

\item{K}{The number of B-spline basis functions, combining degrees of basis functions and the number of knots.
Default is \code{6}, indicating cubic splines.}

\item{degr}{The number of knots in the B-spline base function. Default is \code{3}.}

\item{sigma2}{Error variance. Default is \code{0.1}.}

\item{seed_id}{Seed for random number generation, affecting reproducibility. Default is \code{1000}.}

\item{r1_index}{User-specified sequence of \eqn{r_1} values. 
Default is \code{r1_index}\eqn{ = 1, \dots, \min(\lceil\log(n)\rceil, q)}.
Ignored if \code{isfixedR = 1}.}

\item{r2_index}{User-specified sequence of \eqn{r_2} values. Default sequence spans from 1 to the maximum of \code{K}.
Ignored if \code{isfixedR = 1}.}

\item{r3_index}{User-specified sequence of \eqn{r_3} values. 
Default is \code{r3_index}\eqn{ = 1, \dots, \min(\lceil\log(n)\rceil, q)}.
Ignored if \code{isfixedR = 1}.}

\item{D0}{List of initialization values for each group, with each containing initialized matrices \eqn{S_{(3)}}, \code{A},
\code{B}, and \code{C}. Randomly computed by default, satisfying fixed ranks.}
}
\value{
A list with the following components:
  \item{Y}{Response matrix, dimensions \eqn{n \times q}.}
  \item{X}{Design matrix, dimensions \eqn{n \times p}.}
  \item{f0}{Matrix of true functions, dimensions \eqn{n \times p}.}
  \item{group}{Vector indicating the grouping index of predictors, length \eqn{p}.}
  \item{D0}{Initialized values used in tensor decomposition.}
  \item{...}{Other algorithmic options, dynamically specified.}
}
\description{
Generate scenario II data for a multivariate additive model for multi-view data. This scenario involves
generating data based on sinusoidal transformations of the predictors, incorporating multivariate responses
and a structured group sparsity across tensor decompositions.
}
\examples{
library(comarm)
n <- 200; q <- 5; p <- 100; s <- 3; ng <- 4
group <- rep(1:ng, each = p/ng)
mydata <- marm3.sim.fsin(n, q, p, s, group)

}
\seealso{
\code{\link{marm3.sim.fbs}}
}
