% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marm4.sim.fsin.R
\name{marm4.sim.fsin}
\alias{marm4.sim.fsin}
\title{Generate scenario II data from structural MARM model.}
\usage{
marm4.sim.fsin(
  n,
  q,
  p,
  s,
  group = NULL,
  r10 = 2,
  r20 = 2,
  r30 = 2,
  r40 = 2,
  isfixedR = 0,
  D42 = NULL,
  K = 6,
  degr = 3,
  sigma2 = NULL,
  seed_id = NULL,
  r1_index = NULL,
  r2_index = NULL,
  r3_index = NULL,
  r4_index = NULL,
  D0 = NULL
)
}
\arguments{
\item{n}{Sample size.}

\item{q}{The number of responses, \eqn{q\geq1}.}

\item{p}{The number of covariates, \eqn{p\geq1}.}

\item{s}{The true covariates of each view associating with responses, \eqn{s\geq1}.}

\item{group}{A vector of grouping index for predictors, splitting predictors into groups. By default, it splits 
the predictors into two groups.}

\item{r10}{The first dimension of the tensor. Default is \code{2}.}

\item{r20}{The second dimension of the tensor. Default is \code{2}.}

\item{r30}{The third dimension of the tensor. Default is \code{2}.}

\item{r40}{The fourth dimension of the tensor. Default is \code{2}.}

\item{isfixedR}{Indicates if the tensor ranks are fixed (\code{TRUE} or \code{FALSE}).}

\item{D42}{Mode of unfolding \eqn{D_{(2)}}. Generated randomly by default.}

\item{K}{Number of B-spline basis functions, typically \code{6} for cubic splines.}

\item{degr}{Number of knots in the B-spline base function. Default is \code{3}.}

\item{sigma2}{Error variance. Default is \code{0.1}.}

\item{seed_id}{Seed for random number generation, for reproducibility. Default is \code{1000}.}

\item{r1_index}{User-specified sequence of \eqn{r_1} values, ignored if \code{isfixedR} is \code{TRUE}.}

\item{r2_index}{User-specified sequence of \eqn{r_2} values, ignored if \code{isfixedR} is \code{TRUE}.}

\item{r3_index}{User-specified sequence of \eqn{r_3} values, ignored if \code{isfixedR} is \code{TRUE}.}

\item{r4_index}{User-specified sequence of \eqn{r_4} values, ignored if \code{isfixedR} is \code{TRUE}.}

\item{D0}{List of initialization values for matrices \eqn{S_{(4)}}, \code{A}, \code{B}, \code{C}, and \code{D}.}
}
\value{
A list containing generated data elements:
  \item{Y}{Response matrix, dimensions \eqn{n \times q}.}
  \item{X}{Design matrix, dimensions \eqn{n \times p}.}
  \item{f0}{Matrix of true functions, dimensions \eqn{n \times p}.}
  \item{group}{Vector indicating the grouping index of predictors, length \eqn{p}.}
  \item{D0}{Initialized values used in tensor decomposition.}
  \item{...}{Additional options for the algorithm.}
}
\description{
Generate scenario II data for a structural multivariate additive model for multi-view data (MARM).
This function simulates data based on sinusoidal transformations of predictors, suitable for 
analysis with a structural MARM model using fixed or varying tensor ranks.
}
\examples{
library(comarm)
n <- 200; q <- 5; p <- 100; s <- 3; ng = 4
group <- rep(1:ng, each = p/ng)
mydata <- marm4.sim.fsin(n, q, p, s, group)

}
\seealso{
\code{\link{marm4.sim.fbs}}
}
